Rewrite numeric to double

+++ models/zeroex/arbitrum/zeroex_arbitrum_native_fills.sql
@@ -113,7 +113,7 @@ WITH
                   WHEN mp.symbol = 'WETH' THEN (fills.makerTokenFilledAmount / 1e18) * mp.price
                 ELSE COALESCE((fills.makerTokenFilledAmount / pow(10, mt.decimals))*mp.price,(fills.takerTokenFilledAmount / pow(10, tt.decimals))*tp.price)
               END AS volume_usd
-          , cast(NULL as numeric) AS protocol_fee_paid_eth
+          , cast(NULL as double) AS protocol_fee_paid_eth
       FROM {{ source('zeroex_arbitrum', 'ExchangeProxy_evt_RfqOrderFilled') }} fills
       LEFT JOIN {{ source('prices', 'usd') }} tp ON  tp.blockchain = 'arbitrum' and 
           date_trunc('minute', evt_block_time) = tp.minute 
@@ -173,7 +173,7 @@ WITH
                   WHEN mp.symbol = 'WETH' THEN (fills.makerTokenFilledAmount / 1e18) * mp.price
                 ELSE COALESCE((fills.makerTokenFilledAmount / pow(10, mt.decimals))*mp.price,(fills.takerTokenFilledAmount / pow(10, tt.decimals))*tp.price)
               END AS volume_usd
-          , cast(NULL as numeric) AS protocol_fee_paid_eth
+          , cast(NULL as double) AS protocol_fee_paid_eth
         FROM {{ source('zeroex_arbitrum', 'ExchangeProxy_evt_OtcOrderFilled') }} fills
       LEFT JOIN {{ source('prices', 'usd') }} tp ON tp.blockchain = 'arbitrum'  and 
           date_trunc('minute', evt_block_time) = tp.minute
